gitall/requirements.txt
awsebcli==3.12.4
Babel==2.5.1
blessed==1.14.2
boto3==1.4.7
botocore==1.10.4
cement==2.8.2
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
Django==1.11
django-allauth==0.33.0
django-appconf==1.0.2
django-crispy-forms==1.6.1
django-mediumeditor==1.0.0
django-pagedown==0.1.3
django-phonenumber-field==1.3.0
django-storages==1.6.5
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
futures==3.1.1
idna==2.6
jmespath==0.9.3
oauthlib==2.0.4
olefile==0.44
pathspec==0.5.5
phonenumberslite==8.8.8
Pillow==4.2.1
psycopg2==2.7.3.1
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.3
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.8.0
s3transfer==0.1.11
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.47.0
